requirements.txt
streamlit>=1.28.0
langgraph
langchain-core>=0.1.0
pydantic>=2.5.0
networkx>=3.2.0
python-docx>=1.1.0
pdfplumber>=0.10.0
pandas>=2.1.0
pytest>=7.4.0
openai>=1.12.0
httpx>=0.26.0
pyvis>=0.3.0
